+++ WikiFile.h
@@ -5,7 +5,7 @@ class WikiFile {
     FILE *file;
     const Index &index;
     const WikiName &wikiName;
-    int64_t m_size = 0;
+    size_t m_size = 0;
 public:
     explicit WikiFile(const Index &index);
     std::string decompressChunk(WikiIndexChunk &chunk);
@@ -13,8 +13,8 @@ public:
     std::string decompressChunkByIndex(size_t index);
     int open();
     int close();
-    [[nodiscard]] int64_t size() const {return m_size;}
-    int64_t filePos();
+    [[nodiscard]] size_t size() const {return m_size;}
+    size_t filePos();
     std::string extractTerm(std::string term);
     ~WikiFile();
 };
